chore: support esm/cjs entry points
